use stop

+++ app/Console/Commands/RemoveBuildingCache.php
@@ -79,54 +79,56 @@ class RemoveBuildingCache extends Command
                         $this->table($headers,$buildings);
                      }else{
                         $this->error('No Buildings Found');
-                        break;
+                        $stop = 1;
                      }
                     
             }
             $this->line(PHP_EOL.'Please note that you are only removing the building from the audit'.$audit->id.'. We are NOT removing the building(s) from the property for future audits.'.PHP_EOL);
-            do{
-               
-                $buildingId = $this->anticipate('What is the building id of the building you would like to remove from the audit?',$buildingIds);
+            if($stop == 0){
+                do{
+                   
+                    $buildingId = $this->anticipate('What is the building id of the building you would like to remove from the audit?',$buildingIds);
 
-                if($buildingId){
-                    $check = BuildingInspection::where('building_id',$buildingId)->where('audit_id',$audit->id)->first();
+                    if($buildingId){
+                        $check = BuildingInspection::where('building_id',$buildingId)->where('audit_id',$audit->id)->first();
 
-                    if($check){ 
-                        if($this->confirm('You want to delete '.$check->building_name.'? (This will remove everything including findings associated with this building for this audit).')){
-                            $units = $check->building->units->pluck('id')->to_array();
-                            $check->delete();
-                            // remove inspection items for that building on this audit
-                            AmenityInspection::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
-                            AmenityInspection::where('audit_id',$audit->id)->whereIn('unit_id',$units)->delete();
-                            // remove cached building
-                            CachedBuilding::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
-                            // remove cached building
-                            CachedUnit::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
-                            // remove findings assigned to this building
-                            Finding::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
-                            Finding::where('audit_id',$audit->id)->whereIn('unit_id',$units)->delete();
+                        if($check){ 
+                            if($this->confirm('You want to delete '.$check->building_name.'? (This will remove everything including findings associated with this building for this audit).')){
+                                $units = $check->building->units->pluck('id')->to_array();
+                                $check->delete();
+                                // remove inspection items for that building on this audit
+                                AmenityInspection::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
+                                AmenityInspection::where('audit_id',$audit->id)->whereIn('unit_id',$units)->delete();
+                                // remove cached building
+                                CachedBuilding::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
+                                // remove cached building
+                                CachedUnit::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
+                                // remove findings assigned to this building
+                                Finding::where('audit_id',$audit->id)->where('building_id',$buildingId)->delete();
+                                Finding::where('audit_id',$audit->id)->whereIn('unit_id',$units)->delete();
 
-                            $this->line('Building Removed From Audit.'.PHP_EOL);
-                            $buildings = $audit->building_inspections->pluck('building_id','building_name','address')->toArray();
+                                $this->line('Building Removed From Audit.'.PHP_EOL);
+                                $buildings = $audit->building_inspections->pluck('building_id','building_name','address')->toArray();
+                            }
+
+                        } else {
+                            $this->error(PHP_EOL.'Building Not Found'.PHP_EOL);
+                           
                         }
+                    }
 
+                    if(!$this->confirm('Do you have another Building to delete from this audit?')){
+                        $stop = 1;
                     } else {
-                        $this->error(PHP_EOL.'Building Not Found'.PHP_EOL);
-                       
+                        if($this->confirm('Would you like to see an updated list of the buildings?')){
+                                     
+                                     $this->table($headers,$buildings);
+                                    
+                        }
                     }
-                }
 
-                if(!$this->confirm('Do you have another Building to delete from this audit?')){
-                    $stop = 1;
-                } else {
-                    if($this->confirm('Would you like to see an updated list of the buildings?')){
-                                 
-                                 $this->table($headers,$buildings);
-                                
-                    }
-                }
-
-            }while($stop == 0);
+                }while($stop == 0);
+            }
         }else{
             $this->line(PHP_EOL.'Sorry no audit was selected.');
         }
